refactor(settings): split project backend panel
