+++ ui/src/features/settings/SettingsSectionHeader.tsx
@@ -1,0 +1,50 @@
+import type { ReactNode } from "react";
+
+export function SettingsSectionHeader({
+  title,
+  description,
+  meta,
+  actions,
+}: {
+  title: string;
+  description?: string;
+  meta?: string;
+  actions?: ReactNode;
+}) {
+  return (
+    <div style={{ display: "flex", alignItems: "flex-start", gap: 12, marginBottom: 12 }}>
+      <div style={{ minWidth: 0 }}>
+        <div
+          style={{
+            fontSize: 13,
+            fontWeight: 500,
+            color: "var(--t0)",
+            marginBottom: description ? 3 : 0,
+          }}
+        >
+          {title}
+        </div>
+        {description && (
+          <div style={{ fontSize: 11, color: "var(--t3)", lineHeight: 1.45 }}>{description}</div>
+        )}
+      </div>
+      {meta && (
+        <span
+          style={{
+            fontFamily: "var(--font-mono)",
+            fontSize: 11,
+            color: "var(--t3)",
+            whiteSpace: "nowrap",
+          }}
+        >
+          {meta}
+        </span>
+      )}
+      {actions && (
+        <div style={{ marginLeft: "auto", display: "flex", gap: 8, alignItems: "center" }}>
+          {actions}
+        </div>
+      )}
+    </div>
+  );
+}
